move to utils, move to resources
