+++ src/codec.ts
@@ -8,8 +8,9 @@ import type {
 } from "@waku/interfaces";
 import debug from "debug";
 
+import type { IdentityCredential } from "./identity.js";
 import { RlnMessage, toRLNSignal } from "./message.js";
-import { IdentityCredential, RLNInstance } from "./rln.js";
+import { RLNInstance } from "./rln.js";
 
 const log = debug("waku:rln:encoder");
