+++ src/contract/index.ts
@@ -1,1 +1,2 @@
 export { RLNContract } from "./rln_contract.js";
+export * from "./constants.js";
