+++ src/resources/witness_calculator.d.ts
