+++ src/resources/witness_calculator.js
