+++ src/utils/index.ts
@@ -6,3 +6,4 @@ export {
   zeroPadLE,
 } from "./bytes.js";
 export { sha256, poseidonHash } from "./hash.js";
+export { dateToEpoch, epochIntToBytes, epochBytesToInt } from "./epoch.js";
